change setVolume functions to accept a qint64 rather than a quint32
with a quint32 the GUI code would have to handle overflows, which seems
a bit naughty. instead accept a qint64 (offering the same positive integer
range) but also allowing the context to differentiate the UI code's intent
when the value is a negative integer.
consequently qbound the value between 0 and max_volume before passing it on
to pulseaudio which then uses it as an uint32 again.

this prevents me from blowing my ears out when testing with headphones on
:O

+++ src/context.h
@@ -37,16 +37,16 @@ public:
     void clientCallback(const pa_client_info *info);
     void cardCallback(const pa_card_info *info);
 
-    Q_INVOKABLE void setSinkVolume(quint32 index, quint32 volume);
+    Q_INVOKABLE void setSinkVolume(quint32 index, qint64 volume);
     Q_INVOKABLE void setSinkPort(quint32 index, const QString &portName);
 
-    Q_INVOKABLE void setSinkInputVolume(quint32 index, quint32 volume);
+    Q_INVOKABLE void setSinkInputVolume(quint32 index, qint64 volume);
 #warning terminology is messed up... modelIndex is a QModelIndex, what we have here is a MapIndex...
     Q_INVOKABLE void setSinkInputSinkByModelIndex(quint32 index, int sinkModelIndex);
 
-    Q_INVOKABLE void setSourceVolume(quint32 index, quint32 volume);
+    Q_INVOKABLE void setSourceVolume(quint32 index, qint64 volume);
 
-    Q_INVOKABLE void setSourceOutputVolume(quint32 index, quint32 volume);
+    Q_INVOKABLE void setSourceOutputVolume(quint32 index, qint64 volume);
     Q_INVOKABLE void setSourceOutputSinkByModelIndex(quint32 index, int sourceModelIndex);
 
     Q_INVOKABLE void setCardProfile(quint32 cardIndex, const QString &profileName);
@@ -54,7 +54,7 @@ public:
 private:
     void connectToDaemon();
     template <typename PAFunction>
-    void setGenericVolume(quint32 index, quint32 newVolume,
+    void setGenericVolume(quint32 index, qint64 newVolume,
                           pa_cvolume cVolume, PAFunction pa_set_volume);
 
     void reset();
